requirements.txt
selenium~=4.16.0
pandas~=2.1.4
PyYAML~=6.0.1
requests~=2.31.0
httpx~=0.26.0
SQLAlchemy~=2.0.25
aiohttp~=3.9.1